+++ src/wall_calculation.c
@@ -6,7 +6,7 @@
 /*   By: pchateau <pchateau@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/04 11:15:45 by pchateau          #+#    #+#             */
-/*   Updated: 2025/04/07 13:48:34 by pchateau         ###   ########.fr       */
+/*   Updated: 2025/04/08 17:37:42 by pchateau         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 /**
  * Calculate the wall height perceived with one ray.
  */
-int	calculate_wall_height(int distance_between_player_and_wall)
+float	calculate_wall_height(float distance_between_player_and_wall)
 {
-	int	distance_between_player_and_plane;
+	float	distance_between_player_and_plane;
 	
 	distance_between_player_and_plane = calculate_distance_between_player_and_plane();
-	printf("distance joueur/plan: %d\n", distance_between_player_and_plane);
+	printf("distance joueur/plan: %f\n", distance_between_player_and_plane);
 	return (distance_between_player_and_plane * CUBE_SIZE / distance_between_player_and_wall);
 }
